+++ tests/unit_tests/cx_core/stepper/bounce_stepper_test.py
@@ -1,8 +1,9 @@
+from typing import Literal
+
 import pytest
 from cx_const import StepperDir
 from cx_core.stepper import MinMax
 from cx_core.stepper.bounce_stepper import BounceStepper
-from typing_extensions import Literal
 
 
 @pytest.mark.parametrize(